Consolidate get methods
